backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "flask-smorest>=0.46.2",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.56.0",
    "marshmallow>=4.1.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
